test: 최단 경로 관련 통합테스트 생성
